fmt {projet1}	R Documentation

fmt

Description

La fonction formate une quantite, avec un nombre de decimales specifie, la virgule pour distinguer unites et decimales, l'espace comme separateur de milliers.

Usage

fmt(x, n)

Arguments

x
la quantite a formater

n
le nombre de decimales

Examples

fmt(100, 2)
fmt(1234.56789, 2)

[Package projet1 version 0.1.0 Index]