fmt1Decimale {projet1}	R Documentation

fmt1Decimale

Description

La fonction formate un nombre avec une decimale, la virgule pour distinguer unites et decimales, et l'espace comme separateur de milliers.

Usage

fmt1Decimale(x)

Arguments

x
la quantite a formater

Examples

fmt1Decimale(100)
fmt1Decimale(1234.56789)

[Package projet1 version 0.1.0 Index]