fmt {projet1}	R Documentation

fmt

Description

Formate un nombre

Usage

fmt(x, decimales)

Arguments

x
un nombre

decimales
un entier, le nombre de decimales

Details

La fonction formate avec le nombre de decimales specifie, la virgule pour distinguer unites et decimales, l'espace comme separateur de milliers.

Value

Le nombre formate.

Emploie

La fonction https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/format

Warning

N'utilise pas de notation scientifique

Author(s)

Mage RegoR, mage.regor@email.fr

References

https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/format la fonction format

See Also

https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/print pour afficher dans la console

Other fonctions pour formater: fmt1Decimale()

Examples

fmt2Decimales(100, 2)
fmt2Decimales(1234.56789, 2)

[Package projet1 version 0.1.0 Index]